+++ packages-runtime/wevu-compiler/vitest.config.ts
@@ -3,21 +3,18 @@ import { defineProject } from 'vitest/config'
 import { createProjectCoverage } from '../../vitest.coverage'
 
 export default defineProject({
-  cacheDir: path.resolve(__dirname, './.vite'),
   test: {
     alias: [
       {
-        find: '@',
-        replacement: path.resolve(__dirname, './src'),
+        find: '@/',
+        replacement: `${path.resolve(__dirname, './src')}/`,
       },
     ],
     globals: true,
     testTimeout: 60_000,
-    coverage: createProjectCoverage('packages/wevu', {
+    coverage: createProjectCoverage('packages-runtime/wevu-compiler', {
       clean: false,
-      reporter: ['text', 'lcov'],
       exclude: [
-        'src/index.ts', // pure export barrel, exclude from coverage
         '**/dist/**',
       ],
     }),