+++ packages-runtime/wevu/vitest.config.ts
@@ -3,6 +3,7 @@ import { defineProject } from 'vitest/config'
 import { createProjectCoverage } from '../../vitest.coverage'
 
 export default defineProject({
+  cacheDir: path.resolve(__dirname, './.vite'),
   test: {
     alias: [
       {
@@ -12,9 +13,11 @@ export default defineProject({
     ],
     globals: true,
     testTimeout: 60_000,
-    coverage: createProjectCoverage('packages/weapi', {
+    coverage: createProjectCoverage('packages-runtime/wevu', {
       clean: false,
+      reporter: ['text', 'lcov'],
       exclude: [
+        'src/index.ts', // pure export barrel, exclude from coverage
         '**/dist/**',
       ],
     }),